native-agent/backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
anthropic
supabase
psutil
scikit-learn
weasyprint
httpx
pyinstaller
pydantic-settings
